EUI-2448 refactored after request on PR
