+++ src/noc/store/effects/noc.effects.spec.ts
@@ -6,9 +6,7 @@ import * as fromNocEffects from './noc.effects';
 import { of, throwError } from 'rxjs';
 import { cold, hot } from 'jasmine-marbles';
 import { CheckAnswers, SetAnswers, SetAnswersIncomplete, SetAnswerSubmissionFailure, SetCaseReference, SetCaseRefSubmissionFailure, SetCaseRefValidationFailure, SetQuestions, SetSubmissionFailure, SetSubmissionSuccessApproved, SetSubmissionSuccessPending, SubmitNoc } from '../actions/noc.action';
-import { NocQuestion } from '../../models/noc-question.interface';
-import { NocError } from '../../models/noc-error.interface';
-import { NocAnswer } from '../../models/noc-answer.interface';
+import { NocQuestion , NocError , NocAnswer } from '../../models/';
 
 describe('Noc Effects', () => {
   let actions$;
